backend/requirements.txt
annotated-types
anyio
certifi
charset-normalizer
click
distro
dnspython
ecdsa
email_validator
fastapi
greenlet
h11
httpcore
httpx
idna
iniconfig
jiter
jsonpatch
jsonpointer
langchain
langchain-core
langchain-text-splitters
langchain_openai
langsmith
mongoengine
mongomock
openai
orjson
packaging
passlib
pip
pluggy
pyasn1
pydantic
pydantic_core
pydantic-settings
pymongo
pytest
python-dotenv
python-jose
python-multipart
pytz
PyYAML
redis
requests
requests-toolbelt
rsa
sentinels
six
sniffio
SQLAlchemy
starlette
tenacity
tqdm
typing_extensions
typing-inspection
urllib3
uvicorn
zstandard
graphviz
selenium
webdriver-manager
langchain-anthropic